--- a/encoding/delta/length_byte_array.go
+++ b/encoding/delta/length_byte_array.go
@@ -41,17 +41,16 @@ func (e *LengthByteArrayEncoding) DecodeByteArray(dst []byte, src []byte, offset
 		return dst, offsets, e.wrap(err)
 	}
 
-	if err := validateByteArrayLengths(length.values); err != nil {
-		return dst, offsets, e.wrap(err)
-	}
-
 	if size := len(length.values) + 1; cap(offsets) < size {
 		offsets = make([]uint32, size)
 	} else {
 		offsets = offsets[:size]
 	}
 
-	lastOffset := decodeByteArrayLengths(offsets, length.values)
+	lastOffset, invalidLength := decodeByteArrayLengths(offsets, length.values)
+	if invalidLength != 0 {
+		return dst, offsets, e.wrap(errInvalidNegativeValueLength(int(invalidLength)))
+	}
 	if int(lastOffset) > len(src) {
 		return dst, offsets, e.wrap(errValueLengthOutOfBounds(int(lastOffset), len(src)))
 	}
@@ -66,29 +65,8 @@ func (e *LengthByteArrayEncoding) wrap(err error) error {
 	return err
 }
 
-func validateByteArrayLengths(lengths []int32) error {
-	for _, n := range lengths {
-		if n < 0 {
-			return errInvalidNegativeValueLength(int(n))
-		}
-	}
-	return nil
-}
-
 func encodeByteArrayLengths(lengths []int32, offsets []uint32) {
 	for i := range lengths {
 		lengths[i] = int32(offsets[i+1] - offsets[i])
 	}
-}
-
-func decodeByteArrayLengths(offsets []uint32, lengths []int32) uint32 {
-	lastOffset := uint32(0)
-
-	for i, n := range lengths {
-		offsets[i] = lastOffset
-		lastOffset += uint32(n)
-	}
-
-	offsets[len(lengths)] = lastOffset
-	return lastOffset
 }