--- a/encoding/delta/length_byte_array_purego.go
+++ b/encoding/delta/length_byte_array_purego.go
@@ -1,0 +1,18 @@
+//go:build purego || !amd64
+
+package delta
+
+func decodeByteArrayLengths(offsets []uint32, lengths []int32) (uint32, int32) {
+	lastOffset := uint32(0)
+
+	for i, n := range lengths {
+		if n < 0 {
+			return lastOffset, n
+		}
+		offsets[i] = lastOffset
+		lastOffset += uint32(n)
+	}
+
+	offsets[len(lengths)] = lastOffset
+	return lastOffset, 0
+}